+++ modules/elements/elements/hd-progress/element.php
@@ -16,7 +16,7 @@ return [
 			$metadata->set('script:builder-hd-progress', ['src' => Path::get('./js/hd-progress.js'), 'defer' => true]);
 
             // Don't render element if content fields are empty
-            return $node->props['content'] || ($node->props['max'] && $node->props['stop'] && $node->props['animation_step'] && $node->props['animation_speed']);
+            return Str::length($node->props['content']) || ($node->props['max'] && $node->props['stop'] && $node->props['animation_step'] && $node->props['animation_speed']);
 
 		},