+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
+## 1.1.0 (March 09, 2021)
+
+### Added
+
+- panel style option to timeline item element
+
+### Fixed
+
+- left/right aligned images not covering cards with different heights in flipcard element
+
 ## 1.0.0 (February 20, 2021)
 
 ### Added
 
-  - initial release
+- initial release